Monster health, regen, and split working
Shoot method is in place, called from the enemyAI1 script. @Ewillms2
please make sure the shoot method works correctly and the web texture is
not missing.

The FighterAI is attached to a new spider prefab I created :
bad_spider_FIGHTER. This will have the script already attached to it,
and should exhibit the regen and split behavior so far. (No Regen
behavior will be observable as there is nothing to damage to spiders to
begin with... @bravisc @Ewillms2 please pull ASAP and check out these
changes.