Level Update
An update to the level for homework 8 - still not finished.

With this, I have added more to the level where the basic construction
of the stage and platforming puzzles are finished.  Currently there is
nothing in that big room towards the later half of the level.  I was
thinking a maze of some sorts?  But I really don't know - send me your
thoughts or feel free to just add it.  I also dicked around with the
color of the light in the stage - play with it too to find the "feel"
you guys want.

I also started working on this before your guys most recent pushes, so
it is probably lacking a lot of the things you have worked on.  Please
feel free to add them where they should be/ where you they would fit.

I was thinking in terms of the "artifact" our level has to have, I was
thinking maybe a series of illuminating keys?  There are quite a few
places right now where gates could be placed as progression blockers,
forcing the player to explore the level/maze to find the key.   Let me
know what you think.