Sword now does damage
Can change atkDamage number to increase or decrease damage. Also added
if the enemy is at 0 health he despawns. Would like to add an animation
to that later.  Since we want the user to use the left click to attack
the spiders I have it so the spider must be in contact with the sword
and the user must be left clicking. Added it all the the
EnemyHealthSystem script.