Story Text Management System completed
SanityManager - Updated the screen images to include Text for the
messages, and RawImage for the text background.

TextManager - Script takes the Text and RawImage from SanityManager and
will do the updating of the text on the screen, and the coroutine to
keep on the screen for a certain amount of time.

TextZone - Includes collider for where the text will be activated if the
player walks there. Script takes a string for the message that will be
displayed in this zone. When the player enters the collider, only for
the first time, the method in TextManager will be called, displaying the
text to the player for 10 seconds, then disappearing.