Opening scene and organization
removed "testfor4-1-2016.exe" as I believe it wasn't allowing me to sync

Starting menu and hierarchy clean up

Created an "opening scene" that
gives the pretense for the game and the
setting.  This scene will then
take us to the scene that we're used to.

*Note: while in the editor,
when transitioning from the opening scene to
scene "hmwk8-1" the
lighting will appear to be messed up.  While I
haven't figured out how
to fix the transition in the editor, this isn't
too much of a big deal
as this bug doesn't occur once the game is built.

I also cleaned up the
hierarchy in hmwk8-1 using empty gameobjects as
folders.  Now its much
easier to find and organize things.  As long as
nothing is done to these
folder objects, everything should run and
operate as normal.