Falling dead zones and respawning
Created a basic script that will respawn the player if they fall off the
level.  The player will then respawn at start with half of their current
sanity.

It's fairly easy to set up, just create a plane that covers the level's
fall zones and place an empty child object where you would like the
player to respawn.  The script will compare the y coordinates of both
the fallplane and the player and will the respawn the player if they
fall below it - more info is commented in the script.