Tutorial Scene and Gate Guardians
Tutorial scene with basics created, also modified gateDestroy to have
guardian where that monster needs to die before the gate can be
destroyed.

NOTE: Tutorial still needs to be linked at end to the next scene
(spiders I guess?) and the Opening Scene needs to be linked to the
Tutorial now.