Checkpoint system added to Tutorial/spider level
Added checkpoint system. Needs FallPlane prefab and all of the
Checkpoint prefabs go as children. When using checkpoints make sure to
give them a number of which checkpoint it is under the public variables
or it wont work. Also, set the starting checkpoint in the FallPlane to
be the first Checkpoint object at the beginning of the level.