Added new UI Pictures to SanityManager prefab
Created red version of the UI, seems to look better with the game. The
prefab of SanityManager is updated to have this new UI as well as the
Dialog box and the ClosestEnemy info.

NOTE: The jump level does not have the new UI, etc. as the instance of
the prefab is broken. Please insert new prefab into jump level, whoever
is working on it.