Created Credits Scene
Credits scene created.  Right now it's not linked to anywhere, but once
the jump level is finished, it can be linked there.  At it's conclusion,
it loads our current beginning (our current intro before the spider
level).  Once I've finished a proper title screen, we can have it load
that