Spider tutorial changed, Fire level finished
Spider tutorial text changed to a new color, also the things say "Tip: "
before them as well now. Can be changed by simply clicking the Is
Tutorial check box for a TextZone.

Fire level is finished, including particle systems for the fires (they
deplete Sanity as ProximityEnemy tag), sanity healers, a big room at the
end for the boss, and most importantly, the writers story for the
fire/brother.